move to utils, update role
