+++ app/src/pages/playground/SpanPlaygroundPage.tsx
@@ -7,11 +7,7 @@ import { safelyParseJSON } from "@phoenix/utils/jsonUtils";
 import { spanPlaygroundPageLoaderQuery$data } from "./__generated__/spanPlaygroundPageLoaderQuery.graphql";
 import { Playground } from "./Playground";
 import { llmAttributesSchema } from "./schemas";
-
-type PlaygroundSpan = Extract<
-  spanPlaygroundPageLoaderQuery$data["span"],
-  { __typename: "Span" }
->;
+import { transformSpanAttributesToPlaygroundInstance } from "./utils";
 
 export function SpanPlaygroundPage() {
   const data = useLoaderData() as spanPlaygroundPageLoaderQuery$data;
@@ -29,31 +25,4 @@ export function SpanPlaygroundPage() {
   const playgroundInstance = transformSpanAttributesToPlaygroundInstance(span);
 
   return <Playground instances={[playgroundInstance]} />;
-}
-
-function transformSpanAttributesToPlaygroundInstance(
-  span: PlaygroundSpan
-): PlaygroundInstance {
-  const { json: parsedAttributes, parseError } = safelyParseJSON(
-    span.attributes
-  );
-  if (parseError) {
-    throw new Error("Invalid attributes");
-  }
-  const { data, success } = llmAttributesSchema.safeParse(parsedAttributes);
-  if (!success) {
-    throw new Error("Invalid data");
-  }
-  return {
-    id: 1,
-    input: {
-      variables: {},
-    },
-    template: {
-      __type: "chat",
-      messages: data["llm.input_messages"],
-    },
-    output: data["llm.output_messages"],
-    tools: undefined,
-  };
 }
