+++ app/src/pages/playground/constants.tsx
@@ -1,3 +1,1 @@
 export const NUM_MAX_PLAYGROUND_INSTANCES = 2;
-
-export const DEFAULT_CHAT_ROLE = "system";
